3 neue Get Pfade

+++ VerteilteSystemeProjekt/src/main/java/de/vsys/project/general/data/ForecastNecessaryLockdownDays.java
@@ -11,7 +11,7 @@ public class ForecastNecessaryLockdownDays {
         return forecastNecessaryLockdownDays;
     }
 
-    public void setForecastNecessaryLockdownDays(int forecastNecessaryLockdownDays) {
+    public void setForecastNecessaryLockdownDays(double forecastNecessaryLockdownDays) {
         this.forecastNecessaryLockdownDays = forecastNecessaryLockdownDays;
     }
 }
